Improved sitemap model: method/param/header/cookie
The site map model was inconsistent in terms of identifying a
particulare request, because the request's method was missing.
Also, other useful information that burp provides to its extensions was
not included.

To solve this, the following information was added to the site map model:
* request method
* request parameters
* response headers
* response cookies

For that purpose, the following properties including getters and
setters were added to the class HttpMessage:
* String method
* List<String> headers
* List<Cookie> cookies
* List<Parameter> parameters

The new classes Cookie and Parameter hold the values provided by the
burp interfaces ICookie and IParameter. The enum ParameterType holds the
static byte values of IParameter that represent the type of a parameter.

Resolves #43