Fix for #50
This should resolve the issue described by: https://github.com/vmware/burp-rest-api/issues/50

Cause: Use of AND condition instead of OR condition means that user must be using a free Burp version AND have also not specified a project file. However, this ignores two cases:

1. The user mistakenly states a project file with Burp free edition
2. The user does not state a project file with Burp paid edition

Both of these cases require the creation of an empty projectData array.

Resolution: Changed conditional operator from `&&` to `||`.

I also rearranged the operands so that the `Note:` line flows into the logic of the if-statement. 
"Burp Free does not support project..." => "`if (!burpEdition.equalsIgnoreCase('free')`"